requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
openai
groq
faiss-cpu
sentence-transformers
